Missed changes for club movement
