+++ ippon/permissions_test.py
@@ -15,47 +15,6 @@ import ippon.club.models as cl
 from ippon.serializers import PointSerializer
 
 
-class ClubPermissionsTests(django.test.TestCase):
-    def setUp(self):
-        self.user = User.objects.create(username='user', password='password')
-        self.admin = User.objects.create(username='admin', password='password')
-        self.club = cl.Club.objects.create(
-            name='cn1',
-            webpage='http://cw1.co',
-            description='cd1',
-            city='cc1')
-        self.request = unittest.mock.Mock(user=self.user)
-        self.view = unittest.mock.Mock()
-        self.club_admin = self.club.admins.create(user=self.admin)
-        self.permission = ippon.club.permissisons.IsClubAdminOrReadOnlyClub()
-
-
-class ClubPermissionTestsNotAdmin(ClubPermissionsTests):
-    def setUp(self):
-        super(ClubPermissionTestsNotAdmin, self).setUp()
-
-    def test_permits_when_safe_method(self):
-        self.request.method = 'GET'
-        result = self.permission.has_object_permission(self.request, self.view, self.club)
-        self.assertEqual(result, True)
-
-    def test_doesnt_permit_when_unsafe_method(self):
-        self.request.method = 'PUT'
-        result = self.permission.has_object_permission(self.request, self.view, self.club)
-        self.assertEqual(result, False)
-
-
-class ClubPermissionTestsAdmin(ClubPermissionsTests):
-    def setUp(self):
-        super(ClubPermissionTestsAdmin, self).setUp()
-        self.club.admins.create(user=self.user)
-
-    def test_doesnt_permit_when_unsafe_method(self):
-        self.request.method = 'PUT'
-        result = self.permission.has_object_permission(self.request, self.view, self.club)
-        self.assertEqual(result, True)
-
-
 class PlayerPermissionsTest(django.test.TestCase):
     def setUp(self):
         self.user = User.objects.create(username='user', password='password')
